Reduce the time complexity of the function
Reduce the time complexity of the function by checking for divisors from 2 to sqrt(n) only instead of from 2 to n-1. 

since n = a*b, If both a and b were greater than the square root of n, a*b would be greater than n. So at least one of those factors must be less than or equal to the square root of n, and to check if n is prime, we only need to test for factors less than or equal to the square root.