Group Assignment - Journal app.
Problem Statement 
Reflection is an important part of learning and growth. You’re required to create a journal app Ambitious Alex or any user can use to do just that.

Deliverables
As a user I would like to be able to create a new journal, make a new journal entry, see a list of journal entries made and be able to view my last journal entry only.